move updateStudent() from StudentsLogic to DataMigration
